spam logs with details for compatibility checks
We don't want these in master, but we need a way to let users debug
the (rather common) case where Connect does not produce any solutions,
but nothing is scheduled for compute either because there are only incompatible
planning scenes.

This is the set of output I used to debug such a case on my end.